updated fields in settings; fixed async/await error in /test; renamed components/profile -> components/settings
